backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=23.5
pymongo>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.1.2
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
gunicorn>=21.2.0
